finished cleaning
Functions are now in the seperated files and working
Also took out one function in Vote.js, because it is not usefull for now
